fix(eslint): config will not use eslintrc from this project
There was a problem with the way that eslint fix was applied which
basically resulted in test files in this project inheriting from the
`eslintConfig` from _this project_ (`prettier-eslint`). So I added
`useEslintrc: false` to the config which I believe will fix the issue
(can be overridden)

Also added a test for the README example so this doesn't happen again!