feat(fallback-options): allow passing custom fallbacks if no eslint rule (#73)
* docs(contributing): use yarn instead of npm and fix typo

* feat(fallback-options): allow passing custom fallbacks if no eslint rule

Some users may want to be able to fallback to specific prettier options only in the case that the
rule cannot be inferred from the corresponding ESlint rule. We add a `fallbackPrettierOptions` to
the API to allow this. This is different than `prettierOptions` because those always override the
eslint rule, whereas fallback options are only used if the eslint rule is not found. If a fallback
option is not provided and an eslint rule is not found, the default prettier option is used as was
always the case.

Closes #72