feat(css): setting prettier parser option for CSS files (#91)
Prettier supports the `parser` option to indicate the parser to be used when parsing text rather than a file.
This is necessary because it's difficult to detect the source file type without a file extension.
This fix uses the same logic Prettier does to determine if the parser should be `postcss`.
It also avoids running ESLint on .css, .less, and .scss files.