feat(perf): Remove unfixable rules
ESLint always lint with all rules you provide, even if you call it with the fix command.
It does this to be able to show you information about the rules it could not fix.
We don't display those messages, so we can remove all unfixable rules and allow ESLint to do less work and in turn run faster.