feat: Only use "local" modules for formatting
When we resolve configs for prettier and eslint we now use our own version of the respective packages. This way we don't need to check for backwards-compatibility becuase we know the latest compatible version is available, even if the project using prettier-eslint have an older version of prettier installed.

Because of this we can now lower the version requirements of prettier and eslint.