fix: Work around eslint globals config bug
ESLint have a weird behaviour that I would like to
categorize as a bug, but it's not really a bug, in where
using their API to get/create a config you get the
globals property as an object but their `CLIEngine`
class expects it to be an array.

I now convert the property from an object to an array before using the options.

https://github.com/eslint/eslint/issues/7967