fix: Always use prettier and eslint clostest from filePath or cwd
In an attempt to simplify the config resolution code, especially for prettier,
https://github.com/prettier/prettier-eslint/commit/49b762b57b7e7af3b06bd933050c614a91b6742d
ended up borking eslint's ability to resolve the projects config file.

I do believe this is a short coming in eslint's resolve config logic, as prettier was
able to resolve a project's config file properly in the same scenario.

Even though prettier worked as expected in this scenario I updated it's module
to use the same pattern so we have similar logic thoughout the package.