feat(prettier): Respect editor config
Prettier has the ability to read `.editorconfig` while resolving the rules. These are read automatically when using the prettier CLI.
If the `.editorconfig` file is not found the option is ignored and does not result in error.

This might introduce breaking change (in the author's case a fix, though) for the folks
who have  `.editorconfig` in their folder structure, but it is not in line with prettier config.
There might be a legitimate reason for that, but I cannot think of any.