chore(prettier): Take Prettier 2.4.0's jsxBracketSameLine → bracketSameLine
prettier/prettier@4992d9720, released in 2.4.0, deprecated the
`jsxBracketSameLine` option in favor of a new, more generic option,
`bracketSameLine`. See docs:
  https://prettier.io/docs/en/options.html#bracket-line

This has caused a warning

  jsxBracketSameLine is deprecated.

for people using Prettier 2.4.0+, if their ESLint config leads us to
infer and use a value for jsxBracketSameLine.

So, instead of inferring and using a value for jsxBracketSameLine,
do so for the new bracketSameLine.

In d8bf1e3dc, which we released in 14.0.0, we stopped supporting
Prettier versions older than 2.5.1. That means we don't need to
maintain backward-compatible code for users with Prettier <2.4.0,
which doesn't have bracketSameLine.

See discussion:
  https://github.com/prettier/prettier-eslint-cli/pull/430#issuecomment-1124449040